requirements.txt
numpy
Pillow
tensorflow>=1.5
tabulate
tqdm